Create utils/liquibase.js
