+++ generators/utils.js
@@ -72,8 +72,6 @@ module.exports = {
     packageNameToNamespace,
     stringHashCode,
     RandexpWithFaker,
-    parseLiquibaseChangelogDate,
-    getRecentDateForLiquibase,
     gitExec,
     isGitInstalled,
 };
@@ -818,42 +816,4 @@ function isGitInstalled(callback) {
     const code = gitExec('--version', { trace: false }).code;
     if (callback) callback(code);
     return code === 0;
-}
-
-/*
- * Current faker version is 4.1.0 and was release in 2017
- * It is outdated
- * https://github.com/Marak/faker.js/blob/10bfb9f467b0ac2b8912ffc15690b50ef3244f09/lib/date.js#L73-L96
- * Needed for reproducible builds
- */
-function getRecentDate(days = 1, date = new Date()) {
-    const range = {
-        min: 1000,
-        max: (days || 1) * 24 * 3600 * 1000,
-    };
-
-    let future = date.getTime();
-    future -= faker.random.number(range); // some time from now to N days ago, in milliseconds
-    date.setTime(future);
-
-    return date;
-}
-
-function parseLiquibaseChangelogDate(changelogDate) {
-    if (!changelogDate || changelogDate.length !== 14) {
-        throw new Error('A valid changelogDate is required to parse.');
-    }
-    const formatedDate = `${changelogDate.substring(0, 4)}-${changelogDate.substring(4, 6)}-${changelogDate.substring(
-        6,
-        8
-    )}T${changelogDate.substring(8, 10)}:${changelogDate.substring(10, 12)}:${changelogDate.substring(12, 14)}+00:00`;
-    return new Date(Date.parse(formatedDate));
-}
-
-function getRecentDateForLiquibase(days, changelogDate) {
-    let refDate;
-    if (changelogDate) {
-        refDate = parseLiquibaseChangelogDate(changelogDate);
-    }
-    return getRecentDate(days, refDate);
 }
