+++ test/utils.spec.js
@@ -1,5 +1,3 @@
-const expect = require('chai').expect;
-const faker = require('faker');
 const assert = require('yeoman-assert');
 const helpers = require('yeoman-test');
 const sinon = require('sinon');
@@ -380,54 +378,6 @@ describe('JHipster Utils', () => {
             const isGitInstalled = utils.isGitInstalled(callback);
             assert.strictEqual(isGitInstalled, true);
             sinon.assert.calledOnce(callback);
-        });
-    });
-    describe('::parseLiquibaseChangelogDate', () => {
-        describe('when not passing parameters', () => {
-            it('throws', () => {
-                expect(() => utils.parseLiquibaseChangelogDate()).to.throw(/^A valid changelogDate is required to parse\.$/);
-            });
-        });
-        describe('when passing an invalid changelogDate', () => {
-            it('throws', () => {
-                expect(() => utils.parseLiquibaseChangelogDate('1234')).to.throw(/^A valid changelogDate is required to parse\.$/);
-            });
-        });
-        describe('when passing a valid changelogDate', () => {
-            it('returns a date object', () => {
-                assert(utils.parseLiquibaseChangelogDate('20160208210114') instanceof Date);
-            });
-            it('returns the date', () => {
-                assert.equal(utils.parseLiquibaseChangelogDate('20160208210114').toISOString(), '2016-02-08T21:01:14.000Z');
-            });
-        });
-    });
-    describe('::getRecentDateForLiquibase', () => {
-        beforeEach(() => {
-            faker.seed(47);
-        });
-        describe('when not passing parameters', () => {
-            it('returns a date object', () => {
-                assert(utils.getRecentDateForLiquibase() instanceof Date);
-            });
-        });
-        describe('when passing 1 day', () => {
-            it('returns a date object', () => {
-                assert(utils.getRecentDateForLiquibase(1) instanceof Date);
-            });
-        });
-        describe('when passing 1 day with changelogDate', () => {
-            it('returns a date object', () => {
-                assert(utils.getRecentDateForLiquibase(1, '20160208210114') instanceof Date);
-            });
-            it('returns a recent reproducible date', () => {
-                assert.equal(utils.getRecentDateForLiquibase(1, '20160208210114').toISOString(), '2016-02-08T18:17:47.710Z');
-            });
-        });
-        describe('when passing 10 day with changelogDate', () => {
-            it('returns a recent reproducible date', () => {
-                assert.equal(utils.getRecentDateForLiquibase(10, '20160208210114').toISOString(), '2016-02-07T17:46:59.078Z');
-            });
         });
     });
 });
